Rename RPA.Assistant.dialog_types to RPA.Assistant.types
