+++ packages/assistant/src/RPA/Assistant/__init__.py
@@ -1,1 +1,2 @@
 from .library import Assistant
+from .types import Icon, Options, Result, Size, Location
