+++ packages/assistant/src/RPA/Assistant/utils.py
@@ -2,7 +2,7 @@ from typing import Any, List, Optional, Tuple, Union
 
 from RPA.core.types import is_list_like  # type: ignore
 
-from RPA.Assistant.dialog_types import Element, Options
+from RPA.Assistant.types import Element, Options
 
 
 def to_options(
